temporarily bump up to ruby 2.3.8
i haven't touched this project in 3 years and the last commit back
then ran tests against 2.2 and 2.3.  before upgrading to a more modern
version of ruby this change drops the 2.2 test builds and switches to
the lastest 2.3 to minimize changes to verify things still work as
expected.

also, `rubygems-update` 3.x doesn't support ruby <2.3 so this will sort
out the travis-ci build failure.